requirements.txt
openai
Pillow
TextWrapper